feat: update vc to submit beacon committee selections once per epoch
